update metrics test (#66)

+++ octopus-api/src/main/java/org/metahut/octopus/api/dto/PageResponseDTO.java
@@ -12,12 +12,12 @@ public class PageResponseDTO<T> {
     private Integer pageSize;
 
     @ApiModelProperty(value = "total")
-    private Integer total;
+    private Long total;
 
     @ApiModelProperty(value = "data")
     private List<T> data;
 
-    public static <T> PageResponseDTO of(Integer pageNo, Integer pageSize, Integer total, List<T> data) {
+    public static <T> PageResponseDTO of(Integer pageNo, Integer pageSize, Long total, List<T> data) {
         PageResponseDTO response = new PageResponseDTO<>();
         response.setPageNo(pageNo);
         response.setPageSize(pageSize);
@@ -42,11 +42,11 @@ public class PageResponseDTO<T> {
         this.pageSize = pageSize;
     }
 
-    public Integer getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
